- Moved AppHealthPluginSettings from /platform/settings/handlersettings.go to /plugin/apphealth/settings.go - Moved VMWatchPluginSettings from /platform/settings/handlersettings.go to /plugin/vmwatch/settings.go
