+++ platform/settings/handlersettings.go
@@ -5,116 +5,21 @@ import (
 
 	"github.com/Azure/applicationhealth-extension-linux/pkg/logging"
 	"github.com/Azure/applicationhealth-extension-linux/platform/schema"
+	"github.com/Azure/applicationhealth-extension-linux/plugins/apphealth"
+	"github.com/Azure/applicationhealth-extension-linux/plugins/vmwatch"
 	"github.com/Azure/azure-docker-extension/pkg/vmextension"
 	"github.com/pkg/errors"
 )
 
-var (
-	errTcpMustNotIncludeRequestPath    = errors.New("'requestPath' cannot be specified when using 'tcp' protocol")
-	errTcpConfigurationMustIncludePort = errors.New("'port' must be specified when using 'tcp' protocol")
-	errProbeSettleTimeExceedsThreshold = errors.New("Probe settle time (intervalInSeconds * numberOfProbes) cannot exceed 240 seconds")
-	defaultIntervalInSeconds           = 5
-	defaultNumberOfProbes              = 1
-	maximumProbeSettleTime             = 240
-)
+// Aliases for the plugin settings
+type AppHealthPluginSettings = apphealth.AppHealthPluginSettings
+type VMWatchPluginSettings = vmwatch.VMWatchPluginSettings
+type VMWatchSettings = vmwatch.VMWatchSettings
 
 // HandlerSettings holds the configuration of the extension handler.
 type HandlerSettings struct {
 	publicSettings
 	protectedSettings
-}
-
-func (s *AppHealthPluginSettings) protocol() string {
-	return s.Protocol
-}
-
-func (s *AppHealthPluginSettings) requestPath() string {
-	return s.RequestPath
-}
-
-func (s *AppHealthPluginSettings) port() int {
-	return s.Port
-}
-
-func (s *AppHealthPluginSettings) intervalInSeconds() int {
-	var intervalInSeconds = s.IntervalInSeconds
-	if intervalInSeconds == 0 {
-		return defaultIntervalInSeconds
-	} else {
-		return intervalInSeconds
-	}
-}
-
-func (s *AppHealthPluginSettings) numberOfProbes() int {
-	var numberOfProbes = s.NumberOfProbes
-	if numberOfProbes == 0 {
-		return defaultNumberOfProbes
-	} else {
-		return numberOfProbes
-	}
-}
-
-func (s *AppHealthPluginSettings) gracePeriod() int {
-	var gracePeriod = s.GracePeriod
-	if gracePeriod == 0 {
-		return s.intervalInSeconds() * s.numberOfProbes()
-	} else {
-		return gracePeriod
-	}
-}
-
-func (s *HandlerSettings) VMWatchSettings() *VMWatchSettings {
-	return s.publicSettings.VMWatchSettings
-}
-
-// validate makes logical validation on the handlerSettings which already passed
-// the schema validation.
-func (s AppHealthPluginSettings) validate() error {
-	if s.protocol() == "tcp" && s.port() == 0 {
-		return errTcpConfigurationMustIncludePort
-	}
-
-	if s.protocol() == "tcp" && s.requestPath() != "" {
-		return errTcpMustNotIncludeRequestPath
-	}
-
-	probeSettlingTime := s.intervalInSeconds() * s.numberOfProbes()
-	if probeSettlingTime > maximumProbeSettleTime {
-		return errProbeSettleTimeExceedsThreshold
-	}
-
-	return nil
-}
-
-type vmWatchSignalFilters struct {
-	EnabledTags            []string `json:"enabledTags,array"`
-	DisabledTags           []string `json:"disabledTags,array"`
-	EnabledOptionalSignals []string `json:"enabledOptionalSignals,array"`
-	DisabledSignals        []string `json:"disabledSignals,array"`
-}
-
-type VMWatchSettings struct {
-	Enabled               bool                   `json:"enabled,boolean"`
-	SignalFilters         *vmWatchSignalFilters  `json:"signalFilters"`
-	ParameterOverrides    map[string]interface{} `json:"parameterOverrides,object"`
-	EnvironmentAttributes map[string]interface{} `json:"environmentAttributes,object"`
-	GlobalConfigUrl       string                 `json:"globalConfigUrl"`
-	DisableConfigReader   bool                   `json:"disableConfigReader,boolean"`
-}
-
-// AppHealthPluginSettings holds the configuration of the AppHealth plugin, must match schema
-type AppHealthPluginSettings struct {
-	Protocol          string `json:"protocol"`
-	Port              int    `json:"port,int"`
-	RequestPath       string `json:"requestPath"`
-	IntervalInSeconds int    `json:"intervalInSeconds,int"`
-	NumberOfProbes    int    `json:"numberOfProbes,int"`
-	GracePeriod       int    `json:"gracePeriod,int"`
-}
-
-// VMWatchPluginSettings holds the configuration of VMWatch plugin, must match schema
-type VMWatchPluginSettings struct {
-	VMWatchSettings *VMWatchSettings `json:"vmWatchSettings"`
 }
 
 // publicSettings is the type deserialized from public configuration section of
@@ -155,7 +60,7 @@ func ParseAndValidateSettings(lg logging.Logger, configFolder string) (h Handler
 	lg.Info("parsed configuration json")
 	lg.Info("validating configuration logically")
 
-	if err := h.validate(); err != nil {
+	if err := h.Validate(); err != nil {
 		return h, errors.Wrap(err, "invalid configuration")
 	}
 	lg.Info("validated configuration")
